handle both cases

+++ vllm/ir/ops/layernorm.py
@@ -17,5 +17,5 @@ def rms_norm(
     variance = x_var.pow(2).mean(dim=-1, keepdim=True)
     x = x * torch.rsqrt(variance + epsilon)
     if weight is not None:
-        x = x * weight.to(torch.float32)
+        x = x.to(weight.dtype) * weight
     return x.to(orig_dtype)
